+++ Frends.AzureBlobStorage.ListBlobsInContainer/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [1.1.0] - 2023-02-01
+### Added
+- OAuth2 authentication method.
+
+### Changed
+- Rename: Result.URI to Result.URL.
+
+### Fixed
+- Memory leak fix by unloading assembly context after Task execution.
+- Result.URL returns complete URL to each blob.
+
 ## [1.0.0] - 2022-04-14
 ### Added
 - Initial implementation